Move aimodel/dash_plots/ -> statutil/dash_plots/
